+++ ci/1-build/macos/build.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+
+set -eux
+
+flutter build macos --release --flavor "$FLAVOR"
+
+dmgbuild -s ci/1-build/macos/dmg-settings.py "$FLAVORED_NAME" "$ARTIFACT_FLAVORED".macos-universal.dmg