+++ test/flutter_test_config.dart
@@ -7,7 +7,7 @@ Future<void> testExecutable(FutureOr<void> Function() testMain) async {
   if (Platform.isMacOS) {
     Lz4Codec.libraryPath = 'macos/eslz4-mac64.dylib';
   } else if (Platform.isLinux) {
-    Lz4Codec.libraryPath = 'linux/eslz4-linux64.so';
+    Lz4Codec.libraryPath = 'linux/eslz4-linux-x64.so';
   } else if (Platform.isWindows) {
     Lz4Codec.libraryPath = 'windows/eslz4-win64.dll';
   } else {